LayoutTests/platform/ios-simulator/imported/w3c/web-platform-tests/html/semantics/forms/form-control-infrastructure/form-expected.txt

button

fieldset




object

output





label

label-form

label-form-form2

label-with-control 

label-for 

label-with-progress 

label-with-meter 

 label-for-control-form-in-form

 label-for-control-form

label-in-table	label-in-table 	label-in-table-for	

PASS button.form 
PASS fieldset.form 
PASS input.form 
PASS keygen.form 
PASS object.form 
PASS output.form 
PASS select.form 
PASS textarea.form 
FAIL label.form assert_equals: label.form expected null but got Element node <form id="form">
<p><button id="button">button</button>
<...
FAIL label-form.form assert_equals: label.form expected null but got Element node <form id="form">
<p><button id="button">button</button>
<...
FAIL label-form-form2.form assert_equals: label.form expected null but got Element node <form id="form2"></form>
PASS label-with-control.form 
PASS label-for.form 
FAIL label-with-progress.form assert_equals: Sanity check: label.control.form expected (object) null but got (undefined) undefined
FAIL label-with-meter.form assert_equals: label.form expected null but got Element node <form id="form">
<p><button id="button">button</button>
<...
FAIL label-for-control-form-in-form.form assert_equals: label.form expected Element node <form id="form2"></form> but got Element node <form id="form">
<p><button id="button">button</button>
<...
FAIL label-for-control-form.form assert_equals: label.form expected Element node <form id="form2"></form> but got null
PASS label-in-table.form 
FAIL label-in-table-with-control.form assert_equals: label.form expected Element node <form id="form3"></form> but got null
FAIL label-in-table-for.form assert_equals: label.form expected Element node <form id="form3"></form> but got null
